% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{annotate_links_with_treatments_and_tissues}
\alias{annotate_links_with_treatments_and_tissues}
\title{Step 3 - Context specific annotation and filtering of gene regulatory link predictions}
\usage{
annotate_links_with_treatments_and_tissues(m.lead_support_w_motif.grn,
  m.pvalue_differentialExpression, df.experiment_condition_annotation,
  tb.condition_treatments, tb.condition_tissues, v.conditionGroups,
  v.tissueGroups, th.diffexp = 0.05, th.pval.treatment = 0.05,
  th.pval.tissue = 0.05, th.min.samples = 1,
  s.multipleTestCorrection = "none", b.load = "no")
}
\arguments{
\item{m.grn}{= m.grn,}

\item{l.grn_subnetworks}{= l.grn_subnetworks,}

\item{th.min_number_targets}{= 2,}

\item{th.min_number_MR_targets}{= 2,}

\item{th.pval}{= 0.05}
}
\description{
This function filter...
}
