% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{compute_ensemble_regression_with_montecarlo_based_stability_selection}
\alias{compute_ensemble_regression_with_montecarlo_based_stability_selection}
\title{Step 1 - Gene regulatory network inference using ensemble regression with Monte Carlo based threshold selection}
\usage{

  compute_ensemble_regression_with_montecarlo_based_stability_selection(m.foldChange_differentialExpression,
  df.transcriptionFactorAnnotation, df.geneGroups, seed = 1234,
  importance.measure = "impurity", n.trees = 1000,
  n.lead_method_expression_shuffling = 3, n.bootstrap = 100,
  n.stepsLARS = 5, n.cpus = 5)
}
\arguments{
\item{k}{(default k="sqrt")}
}
\description{
Step 1 - Gene regulatory network inference using ensemble regression with Monte Carlo based threshold selection
}
