% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{do_master_regulator_hierarchy_inference}
\alias{do_master_regulator_hierarchy_inference}
\title{Step 4 - Master regulator hierarchy inference}
\usage{
do_master_regulator_hierarchy_inference(m.grn, l.grn_subnetworks,
  df.transcriptionFactorAnnotation, df.geneGroups, tb.geneGroups,
  v.geneGroups, l.geneGroups, th.min_number_targets = 2,
  th.min_number_MR_targets = 2, th.pval = 0.05,
  foldername.results = "results/")
}
\arguments{
\item{m.grn}{= m.grn,}

\item{l.grn_subnetworks}{= l.grn_subnetworks,}

\item{th.min_number_targets}{= 2,}

\item{th.min_number_MR_targets}{= 2,}

\item{th.pval}{= 0.05}
}
\description{
This function identifies the master regulator hierarchy
}
