% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{format_results}
\alias{format_results}
\title{format_results}
\usage{
format_results(l.grn_subnetworks,
  tb.condition_tissue_differentialExpression, l.Hierarchy,
  l.Hierarchy_tfs_per_tier, l.Hierarchy_nb_tfs_per_tier,
  l.df.masterRegulatorHierarchy, v.number_tiers, m.MR_vs_conditions,
  l.MR_vs_geneGroups_given_condition,
  number_of_conditions_per_master_regulator, tb.condition_treatments,
  tb.condition_tissues, df.transcriptionFactorAnnotation, df.geneGroups,
  tb.geneGroups, v.geneGroups, l.geneGroups, th.pval = 0.05,
  foldername.results = "results/")
}
\description{
This function formats all results
}
