% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METACLUSTER_functions.R
\name{load_gene_cluster_data}
\alias{load_gene_cluster_data}
\title{Load gene cluster Function}
\usage{
load_gene_cluster_data(filename.geneCluster = "",
  input_format = "PCF2017")
}
\arguments{
\item{filename.geneCluster}{includes gene cluster filenames as well as the input}

\item{input_format}{format of plant cluster finder (PCF2017) or custom}
}
\description{
This function loads gene cluster information
}
\examples{
load_gene_cluster_data()
}
