% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{load_lead_support_grn}
\alias{load_lead_support_grn}
\title{load grns}
\usage{
load_lead_support_grn(df.transcriptionFactorAnnotation, df.geneGroups,
  th.lead_grn_method = 0.95, n.lead_method_expression_shuffling = 3,
  th.support_grn_methods = 0.95, n.grnSupport = 1)
}
\arguments{
\item{df.transcriptionFactorAnnotation}{=l.data$df.transcriptionFactorAnnotation,}

\item{df.geneGroups}{= v.genes,}

\item{th.lead_grn_method}{= 0.95,}

\item{n.lead_method_expression_shuffling}{= 3}

\item{th.support_grn_methods}{= 0.95,}

\item{ngrnSupport}{= 1}
}
\description{
This function loads a computed grns
}
