% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METACLUSTER_functions.R
\name{save_pheatmap_pdf}
\alias{save_pheatmap_pdf}
\title{save heatmapt}
\usage{
save_pheatmap_pdf(x, filename, width = 7, height = 7)
}
\arguments{
\item{x}{matrix}
}
\description{
This function saves a heatmap plot
}
\examples{
save_pheatmap_pdf()
}
