% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{stabilitySelection}
\alias{stabilitySelection}
\title{stabilitySelectionProcedure}
\usage{
stabilitySelection(x, y, nbootstrap = 500, nsteps = 5, alpha = 0.2,
  plotme = FALSE)
}
\description{
stabilitySelectionProcedure
}
