% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{transcriptionFactorBindingInference}
\alias{transcriptionFactorBindingInference}
\title{Step 2 - Transcription factor direct target promoter binding based filtering of gene regulatory link predictions}
\usage{
transcriptionFactorBindingInference(m.grn, file.TF_to_Motif_IDs = "",
  file.TFBS_motifs = "", file.promoterSeq = "", file.geneSeq = "",
  th.pre_tss = 1000, th.post_tss = 200, th.min.score.motif = "80\%",
  genome_nucleotide_distribution = c(0.3253439, 0.1746561, 0.1746561,
  0.3253439), th.pval.known_motifs = 0.05, n.cpus = 2, b.load = "no")
}
\arguments{
\item{m.grn}{gene regulatory network}
}
\description{
Step 2 - Transcription factor direct target promoter binding based filtering of gene regulatory link predictions
}
\examples{
install_and_load_libraries()
}
