% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METACLUSTER_functions.R
\name{evaluate_and_store_results}
\alias{evaluate_and_store_results}
\title{A result statistics}
\usage{

  evaluate_and_store_results(df.cluster_annotations = df.cluster_annotations,
  m.functionality = m.functionality, foldername.results = "results/",
  width.heatmap = 7, height.heatmap = 7)
}
\arguments{
\item{df.cluster_annotations}{inferred cluster condition annotations}

\item{m.functionality}{gene cluster acitivity heatmap}
}
\description{
This function analysis the results
}
\examples{
cat_function()
}
