% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METACLUSTER_functions.R
\name{evaluate_and_store_results}
\alias{evaluate_and_store_results}
\title{A result statistics}
\usage{

  evaluate_and_store_results(df.cluster_annotations = df.cluster_annotations,
  df.experiment_condition_annotation = df.experiment_condition_annotation,
  m.functionality = m.functionality, v.gc_validated = c("C628_3",
  "C463_3", "C615_3", "C641_3"), heatmap_width = 10,
  heatmap_height = 6, foldername.results = "results/")
}
\arguments{
\item{df.cluster_annotations}{inferred cluster condition annotations}

\item{m.functionality}{gene cluster acitivity heatmap}
}
\description{
This function analysis the results
}
\examples{
cat_function()
}
