% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METACLUSTER_functions.R
\name{load_datasets}
\alias{load_datasets}
\title{Load dataset function}
\usage{
load_datasets(input_format = "PCF2017", filename.genes = "",
  filename.experiment_series_ids = "", filename.condition_groups = "",
  filename.geneCluster = "",
  filename.foldChange_differentialExpression = "",
  filename.pvalue_differentialExpression = "",
  filename.experiment_condition_tissue_annotation = "")
}
\arguments{
\item{input_format}{"custom" or "PCF2017" (default = "PCF2017")}

\item{filename.genes}{genes (rows of the expression datasets)}

\item{filename.experiment_series_ids}{experimental datasets (columns of the expression datasets)}

\item{filename.condition_groups}{treatment and tissue to condition maps}

\item{filename.geneCluster}{filename gene clusters}

\item{filename.foldChange_differentialExpression}{differential expression data (fold changes)}

\item{filename.pvalue_differentialExpression}{differential expression data (p-values)}

\item{filename.experiment_condition_tissue_annotation}{experiment to treatment and tissue annotation}
}
\description{
This function loads a datasets
}
\examples{
load_datasets()
}
